implemented autosave functionality
autosave is timer based.
changed the way the objects are serialized to not reset data on serialize because of autosave functionality.
added search/filter capability to master lists for request and sequences.
removed save button.
added first try to parse complextype and add this data to the entitytypes. still work in progress. has to be a recursive solution.

Signed-off-by: Kuesters, Elmar <elmar.kuesters@sap.com>